Api class structure (#42)
* add manage collaborators

* fix PR comments

* Refactor permision-based visibility of elements

* Add live data for project and collaborators details on /manage/*

* Add auth button and script

* authenticate projectsHighlights and projects listing

* get authorization from TPEN_User

* Bring in development + suggestions (#38)

* logout redirect hotfix

* logout redirect hotfix

* Tpen class (#36)

* Update index.mjs

bad merge added local env

* Create index.mjs

* TPEN class with authentication

* match docs to class

* swap out auth

* adding expiry

* cleaning up the auth

* Add tpen class auth to transcription (#37)

* replacing URLsafed chars with base64 equiv

* working with new auth

* removing excess

* error handling

* moving the file to a Class

* Starting here.

* auth is for TPEN

* userMessages

* notes

* get User out of TPEN

* collaborators updated.

---------


* relocated User class to /api

* out of date local

* migrate TPEN modules

* Project services tie-in

* fix import path error

* quick fix

* use the new public-login component for the auth-button.

* new login button on /projects

* listing

* rough - in progress

* up and running

* Update index.html

* helpful cheat

* tidying references in imports

* relocation

* trying to bring in changes

* cleaning up merge

* bad User error fixed

* contributer > collaborator

* working if you want it.

services isn't passing displayName back yet.

* null check

* This is definitely wrong

* touch up loading behavior

* Watch for user-id changes

* what's happening. This is in the gitignore already

* Update User.mjs

* cleanup after merge

* add TPEN, await more

* import missing no more

* mh

* no TPEN_USER

* fix add/remove contributor

* hotfix remove error

* update UI after removal

* Manage roles (#43)

* add manage roles components

* add role management services

* add role management controls

* add manage roles controls

* clean up

* quickfix

* Update index.mjs

* member management

* Update project.mjs

* moved Project

* Update index.mjs

* Update events.mjs

* TPEN module singletoned

* Update index.mjs

* Just documentation/comments stuff

* Just documentation/comments stuff

---------
