+++ TPEN/events.mjs
@@ -1,23 +1,7 @@
-class EventDispatcher extends EventTarget {
-    constructor() {
-        super()
-    }
+/**
+ * @deprecated: use and update the file in /api instead
+ */
+console.warn('Deprecated: use and update the file in /api instead')
 
-    // Method to add an event listener
-    on(event, listener) {
-        this.addEventListener(event, listener)
-    }
-
-    // Method to remove an event listener
-    off(event, listener) {
-        this.removeEventListener(event, listener)
-    }
-
-    // Method to dispatch an event
-    dispatch(event, detail = {}) {
-        this.dispatchEvent(new CustomEvent(event, { detail }))
-    }
-}
-
-// Export a shared instance of EventDispatcher
-export const eventDispatcher = new EventDispatcher()
+import { eventDispatcher } from '../api/events.mjs'
+export default eventDispatcher
