+++ components/user-profile/index.html
@@ -6,9 +6,9 @@
     <meta charset="UTF-8">
     <link href="/favicon.ico" rel="icon" type="image/x-icon" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="../../TPEN/index.mjs" type="module"></script>
+    <script src="../../api/TPEN.mjs" type="module"></script>
     <script src="index.mjs" type="module"></script>
-    <script src="../../TPEN/events.mjs" type="module"></script>
+    <script src="../../api/events.mjs" type="module"></script>
 </head>
 
 <body>
