+++ manage/index.html
@@ -10,7 +10,7 @@
   <link rel="stylesheet" href="../css/manage/collaboration.css" />
   <script type="module" src="./index.mjs"></script>
   <script type="module" src="../components/public-login/index.js"></script>
-  <script type="module" src="../TPEN/index.mjs"></script>
+  <script type="module" src="../api/TPEN.mjs"></script>
 </head>
 
 <body>
