+++ projects/index.mjs
@@ -1,12 +1,10 @@
-import  User  from "../User/index.mjs"
-import checkUserAuthentication from "../utilities/checkUserAuthentication.mjs"
-import getHash from "../utilities/getHash.mjs"
+import  User  from "../api/User.mjs"
+import TPEN from "../api/TPEN.mjs"
 
 document.addEventListener("DOMContentLoaded", async () => {
-    const TPEN_USER =  await checkUserAuthentication()
-    let token = TPEN_USER?.authentication
-    let userID = getHash(TPEN_USER.agent)
-    const userObj = new User(userID)
-    userObj.authentication = token
-    userObj.renderProjects("projects-container")
+    const projectsList = document.getElementById("projects-container")
+    TPEN.attachAuthentication(projectsList)
+    const userId = projectsList.getAttribute("tpen-user-id")
+    const user = new User(userId)
+    user.renderProjects("projects-container")
 })
