+++ utilities/getActiveProject.mjs
@@ -1,6 +1,8 @@
-// @deprecated - use TPEN.#activeProject instead
+/**
+ * @deprecated This module has been deprecated. Please use TPEN.#activeProject instead
+ */
 
-import Project from "../api/project.mjs"
+import Project from "../api/Project.mjs"
 
 export default async function getActiveProject() {
     const URLParams = new URLSearchParams(window.location.search)
