+++ utilities/renderRoles.mjs
@@ -1,3 +1,13 @@
+// export default function renderRoles(roles) {
+//     if (roles.includes("OWNER")) {
+//         return "owner"
+//     } else if (roles.includes("LEADER")) {
+//         return "leader"
+//     } else {
+//         return roles.join(", ").toLowerCase().replaceAll("_", " ")
+//     }
+// }
+
 export default function renderRoles(roles) {
     const defaultRoles = ["OWNER", "LEADER", "CONTRIBUTOR", "VIEWER"]
